documents after i launched the program
it creates a knowledge_log every time i use the app
doesnt always record useful information though...
spectra_history
im not quite sure what that does, i guess its sorta like patch notes?
but the weird thing is, is that its a UI element, and its like a drop down menu? idk. Its kind of weird, I think it should be a subsection of the docs tab if anything, but also we dont need to see everything it has in there right? im not certain. if we're going to have it exporting some kind of history, it should be useful to the next ai agent that works on it. so they can know what kind of issues we may have ran into, or what we did.
we do have a set of documentation for bug logging though, and all kinds of other stuff. the agents need to utilize this better.