additional docs, notes for codex to review
created atlas entries (as of now there are 33 documents)
added readmes for various folders.
deleted all previous sample data
uploaded new sample data in the form of lamps. i will upload more as i continue, and this sample data will update as i add files to it. (ie the folder display for samples updates and stays consistent)
i will upload other forms of sample data, not just lamp data

revised some wording in the MASTER_PROMPT.md.
im not sure if any agent has been refering to this as of late, but i cut out some stuff that could cause issues.

deleted doc\atlas\brains.md
and created "doc\brains\" folder instead for individual entry storage

added some /changes to some of the atlas chapters, cause a lot of it is inaccurate, and should be revised by future agents to make our atlas accurately reflect our work as we do it, each step of the way.

we should also revise how we do the dataset/library areas. it has become a cluttered buggy mess. loading cached files does not work. there is now way to clear all data. we should break up some grouping or sections for the displayed data, maybe by instrument, or type, or something to at least categorize it in some way as it is being displayed, so we can easily differentiate between data sets without losing accuracy.

for some reason the agents have not installed modules for astrophy and astroquery and other necessary programs. check all install paths to make sure we have the right things we need, and that this is updated as we add in more features.
this will show up as a bug in the fetch remote data option. when the user clicks on a database to fetch from (ie NIST) there is nothing but a blank, inactive drop down box. there is text displayed in the window noting that fetching cannot be performed without the required tools. this should be fixed so we have a working fetching database.

sample data shouldnt be displayed right away, it should be in its own section in the data sets section, wherever it needs to be.

we should review how we are displaying/grouping/organizing tabs and docks/modules/boxes/graphs in our program. as of now it is manageable, but we should make the user interface more friendly.
we should revisit the keybindings, and remake/remap them with better more logical bindings to more features.
We should revisit how we are using toggles, displays, functions, all of that jazz.
make everything make sense in the ui, without over complicating it.
the option to change options on the mouse is nice, but a lot of functionality is confined only to the mouse, this should be revisited to better improve ui functionality.

## copy pasting last comment from previous manual push, so chat can see it ##

it creates a knowledge_log every time i use the app
doesnt always record useful information though...
spectra_history
im not quite sure what that does, i guess its sorta like patch notes?
but the weird thing is, is that its a UI element, and its like a drop down menu? idk. Its kind of weird, I think it should be a subsection of the docs tab if anything, but also we dont need to see everything it has in there right? im not certain. if we're going to have it exporting some kind of history, it should be useful to the next ai agent that works on it. so they can know what kind of issues we may have ran into, or what we did.
we do have a set of documentation for bug logging though, and all kinds of other stuff. the agents need to utilize this better.