feat(palette): add palette controls for uniform color selection and persistence
ADDED ACUTAL CLEAN UP FILE

going to try to fix the cluttered mess that is this repo. and add in comments for future use. in addition to clean/logical folder and file structure with self updating references and file locations.

gonna try to get documentation cleaned up as well cause a lot of it is out of date.

probably going to have to go file by file folder by folder to manually verify each

So im taking this to a clean up branch